backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-aiplatform==1.51.0
google-api-python-client==2.126.0
google-auth-oauthlib==1.2.0
google-cloud-speech==2.26.0
google-cloud-pubsub==2.21.0
google-cloud-secret-manager==2.20.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.4.4
mypy==1.10.0
httpx==0.27.0
python-multipart==0.0.9
psycopg2-binary
alembic